Misc PR adjustments; refactor + update unit test.
