+++ src/models/enums/fulfillmentType.ts
@@ -1,0 +1,4 @@
+export enum FulfillmentType {
+  CollectionByBuyer = 'Collection by buyer',
+  DeliveredToBuyer = 'Delivered to buyer'
+}
