+++ test/mockData.json
@@ -43,6 +43,15 @@
       "search_map_center": { 
         "type": "Point",
         "coordinates": [-83.856077, 50.848447] 
+      },
+      "search_filters": {
+        "include_active_sellers": true,
+        "include_inactive_sellers": true,
+        "include_test_sellers": true,
+        "include_trust_level_100": true,
+        "include_trust_level_80": true,
+        "include_trust_level_50": true,
+        "include_trust_level_0": true
       }
     },
     {
@@ -56,6 +65,15 @@
       "search_map_center": { 
         "type": "Point",
         "coordinates": [-73.856077, 40.848447] 
+      },
+      "search_filters": {
+        "include_active_sellers": true,
+        "include_inactive_sellers": false,
+        "include_test_sellers": true,
+        "include_trust_level_100": true,
+        "include_trust_level_80": true,
+        "include_trust_level_50": false,
+        "include_trust_level_0": false
       }
     },
     {
@@ -90,8 +108,8 @@
       "description": "Test Seller 1 Description",
       "seller_type": "testSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [-74.0060, 40.7128]
+        "type": "Point",
+        "coordinates": [-74.0060, 40.7128]
       }
     },
     {
@@ -100,8 +118,8 @@
       "description": "Test Vendor 2 Description",
       "seller_type": "inactiveSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [-118.2437, 34.0522]
+        "type": "Point",
+        "coordinates": [-118.2437, 34.0522]
       }
     },
     {
@@ -110,8 +128,8 @@
       "description": "Test Vendor 3 Description",
       "seller_type": "activeSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [-87.6298, 41.8781]
+        "type": "Point",
+        "coordinates": [-87.6298, 41.8781]
       }
     },
     {
@@ -120,8 +138,8 @@
       "description": "Test Seller 4 Description",
       "seller_type": "inactiveSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [-122.4194, 37.7749]
+        "type": "Point",
+        "coordinates": [-122.4194, 37.7749]
       }
     },
     {
@@ -130,8 +148,8 @@
       "description": "Test Vendor 5 Description",
       "seller_type": "testSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [-95.3698, 29.7604]
+        "type": "Point",
+        "coordinates": [-95.3698, 29.7604]
       },
       "fulfillment_method": "Collection by buyer",
       "fulfillment_description": "Test Vendor 5 Fulfillment Description"
@@ -143,8 +161,8 @@
       "address": "Sanctioned Region Cuba",
       "seller_type": "testSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [-84.3829, 22.132]
+        "type": "Point",
+        "coordinates": [-84.3829, 22.132]
       }
     },
     {
@@ -154,8 +172,8 @@
       "address": "Sanctioned Region Iran",
       "seller_type": "testSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [46.7324, 37.4585]
+        "type": "Point",
+        "coordinates": [46.7324, 37.4585]
       }
     },
     {
@@ -165,8 +183,8 @@
       "address": "Sanctioned Region North Korea",
       "seller_type": "testSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [128.7934, 41.1544]
+        "type": "Point",
+        "coordinates": [128.7934, 41.1544]
       }
     },
     {
@@ -176,8 +194,8 @@
       "address": "Sanctioned Region Syria",
       "seller_type": "testSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [38.4684, 35.0528]
+        "type": "Point",
+        "coordinates": [38.4684, 35.0528]
       }
     },
     {
@@ -187,8 +205,8 @@
       "address": "Sanctioned Region Republic of Crimea",
       "seller_type": "testSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [34.0107, 45.4699]
+        "type": "Point",
+        "coordinates": [34.0107, 45.4699]
       }
     },
     {
@@ -198,8 +216,8 @@
       "address": "Sanctioned Region Donetsk Oblast",
       "seller_type": "testSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [36.7852, 48.7462]
+        "type": "Point",
+        "coordinates": [36.7852, 48.7462]
       }
     },
     {
@@ -209,8 +227,8 @@
       "address": "Sanctioned Region Luhansk Oblast",
       "seller_type": "testSeller",
       "sell_map_center": {
-          "type": "Point",
-          "coordinates": [39.9930, 49.5582]
+        "type": "Point",
+        "coordinates": [39.9930, 49.5582]
       }
     }
   ],
